# client-sdks/manager/rust/openapi-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alien Manager API",
    "description": "Control plane for Alien applications. Manages deployments, releases, commands, and telemetry.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/commands/{command_id}/response": {
      "put": {
        "tags": [
          "deployments"
        ],
        "summary": "Submit response from deployment",
        "operationId": "submit_response",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response submitted successfully"
          },
          "400": {
            "description": "Invalid command or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Command not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BodySpec": {
        "oneOf": [
          {
            "type": "object",
            "description": "Inline base64-encoded body",
            "required": [
              "inlineBase64",
              "mode"
            ],
            "properties": {
              "inlineBase64": {
                "type": "string"
              },
              "mode": {
                "type": "string",
                "enum": [
                  "inline"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Storage-backed body",
            "required": [
              "mode"
            ],
            "properties": {
              "mode": {
                "type": "string",
                "enum": [
                  "storage"
                ]
              },
              "size": {
                "type": "integer",
                "format": "int64",
                "description": "Size of the body in bytes",
                "minimum": 0,
                "nullable": true
              },
              "storageGetRequest": {
                "$ref": "#/components/schemas/PresignedRequest",
                "description": "Pre-signed request for retrieving the body (for deployments)",
                "nullable": true
              },
              "storagePutUsed": {
                "type": "boolean",
                "description": "Indicates storage upload was used for response submission",
                "nullable": true
              }
            }
          }
        ],
        "description": "Body specification supporting inline and storage modes"
      },
...
      "CommandResponse": {
        "oneOf": [
          {
            "type": "object",
            "description": "Command executed successfully",
            "required": [
              "response",
              "status"
            ],
            "properties": {
              "response": {
                "$ref": "#/components/schemas/BodySpec",
                "description": "Response data (JSON, can be large)"
              },
              "status": {
                "type": "string",
                "enum": [
                  "success"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Command failed with an error",
            "required": [
              "code",
              "message",
              "status"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code"
              },
              "details": {
                "type": "string",
                "description": "Optional additional details",
                "nullable": true
              },
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "status": {
                "type": "string",
                "enum": [
                  "error"
                ]
              }
            }
          }
        ],
        "description": "Command response from deployment"
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response wrapper for API endpoints",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LocalOperation": {
        "type": "string",
        "description": "Local filesystem operations",
        "enum": [
          "put",
          "get",
          "delete"
        ]
      },
...
      "PresignedOperation": {
        "type": "string",
        "description": "The type of operation a presigned request performs",
        "enum": [
          "put",
          "get",
          "delete"
        ]
      },
      "PresignedRequest": {
        "type": "object",
        "description": "A presigned request that can be serialized, stored, and executed later.\nHides implementation details for different storage backends.",
        "required": [
          "backend",
          "expiration",
          "operation",
          "path"
        ],
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/PresignedRequestBackend",
            "description": "The storage backend this request targets"
          },
          "expiration": {
            "type": "string",
            "format": "date-time",
            "description": "When this presigned request expires"
          },
          "operation": {
            "$ref": "#/components/schemas/PresignedOperation",
            "description": "The operation this request performs"
          },
          "path": {
            "type": "string",
            "description": "The path this request operates on"
          }
        }
      },
      "PresignedRequestBackend": {
        "oneOf": [
          {
            "type": "object",
            "description": "HTTP-based request (AWS S3, GCP GCS, Azure Blob)",
            "required": [
              "url",
              "method",
              "headers",
              "type"
            ],
            "properties": {
              "headers": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "method": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "http"
                ]
              },
              "url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "Local filesystem operation",
            "required": [
              "filePath",
              "operation",
              "type"
            ],
            "properties": {
              "filePath": {
                "type": "string"
              },
              "operation": {
                "$ref": "#/components/schemas/LocalOperation"
              },
              "type": {
                "type": "string",
                "enum": [
                  "local"
                ]
              }
            }
          }
        ],
        "description": "Storage backend representation for different presigned request types"
      },
...
      "SubmitResponseRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommandResponse",
            "description": "The command response"
          }
        ],
        "description": "Request to submit a command response (from deployment)"
      },
...
    }
  },
...
}